// vue/src/app/router/routes.enum.ts
enum ROUTE_NAMES {
  HOME = 'home',

  APPLICATIONS = 'applications',

  GAMES = 'games',
  GAMES_LIST = 'games-list',
  GAMES_SINGLE = 'games-single',

  ADMINISTRATION = 'administration',

  ADMIN_USERS = 'user-administration',
  ADMIN_USER_LIST = 'user-list-administration',

  ADMIN_GAMES = 'game-library-administration',
  ADMIN_GAMES_LIST = 'game-library-list-administration',
  ADMIN_GAMES_PLATFORMS_LIST = 'game-platform-list-administration',

  SETTINGS = 'settings',
  SECURITY = 'security',
  PROFILE = 'profile',

  ACCOUNT_RECOVERY = 'account-recovery',
  PASSWORD_RESET = 'password-reset'
}

export { ROUTE_NAMES }